scripts/diameter measurer/pipeline_codes/requirements.txt
matplotlib==3.1.2
numpy==1.17.4
opencv_contrib_python==4.5.5.62
openpyxl==3.1.2
Pillow==9.5.0
plantcv==3.14.2
pykuwahara==0.3.2
pytesseract==0.3.10
scikit_image==0.18.3
skimage==0.0
tensorflow==2.12.0

Python 3.8.10
pip 23.1.2 from /home/marilin/.local/lib/python3.8/site-packages/pip (python 3.8)

